header reorg, rename variables, reorder debug statements
