+++ src/legacy/finufft1d.cpp
@@ -45,9 +45,9 @@ int finufft1d1(BIGINT nj,FLT* xj,CPX* cj,int iflag,FLT eps,BIGINT ms,
 
   BIGINT n_modes[3] = {ms,1,1};
   int n_dims = 1;
-  int n_vecs = 1;
+  int n_transf = 1;
   finufft_type type = type1;
-  int ier = invokeGuruInterface(n_dims, type, n_vecs, nj, xj, NULL, NULL, cj,
+  int ier = invokeGuruInterface(n_dims, type, n_transf, nj, xj, NULL, NULL, cj,
 				iflag, eps, n_modes, fk, opts);
  
   return ier;
@@ -91,9 +91,9 @@ int finufft1d2(BIGINT nj,FLT* xj,CPX* cj,int iflag,FLT eps,BIGINT ms,
 
   BIGINT n_modes[3] = {ms,1,1};
   int n_dims = 1;
-  int n_vecs = 1;
+  int n_transf = 1;
   finufft_type type = type2;
-  int ier = invokeGuruInterface(n_dims, type, n_vecs, nj, xj, NULL, NULL, cj,
+  int ier = invokeGuruInterface(n_dims, type, n_transf, nj, xj, NULL, NULL, cj,
 				iflag, eps, n_modes, fk, opts);
 
   return ier;
