+++ src/legacy/finufft2d.cpp
@@ -55,9 +55,9 @@ int finufft2d1(BIGINT nj,FLT* xj,FLT *yj,CPX* cj,int iflag,
 
   BIGINT n_modes[3] = {ms,mt,1};
   int n_dims = 2;
-  int n_vecs = 1;
+  int n_transf = 1;
   finufft_type type = type1;
-  int ier = invokeGuruInterface(n_dims, type, n_vecs, nj, xj, yj, NULL, cj, iflag,
+  int ier = invokeGuruInterface(n_dims, type, n_transf, nj, xj, yj, NULL, cj, iflag,
 		      eps, n_modes, fk, opts);
 
   
@@ -158,9 +158,9 @@ int finufft2d2(BIGINT nj,FLT* xj,FLT *yj,CPX* cj,int iflag,FLT eps,
   
   BIGINT n_modes[3] = {ms,mt,0};
   int n_dims = 2;
-  int n_vecs = 1;
+  int n_transf = 1;
   finufft_type type = type2;
-  int ier = invokeGuruInterface(n_dims, type, n_vecs, nj, xj, yj, NULL, cj, iflag,
+  int ier = invokeGuruInterface(n_dims, type, n_transf, nj, xj, yj, NULL, cj, iflag,
 		      eps, n_modes, fk, opts);
