+++ src/legacy/invokeGuru.cpp
@@ -1,7 +1,7 @@
 #include <invokeGuru.h>
 
 
-int invokeGuruInterface(int n_dims, finufft_type type, int n_vecs, BIGINT nj, FLT* xj,FLT *yj, FLT *zj, CPX* cj,int iflag,
+int invokeGuruInterface(int n_dims, finufft_type type, int n_transf, BIGINT nj, FLT* xj,FLT *yj, FLT *zj, CPX* cj,int iflag,
 			FLT eps, BIGINT *n_modes, CPX* fk, nufft_opts opts){
 
 
@@ -9,7 +9,7 @@ int invokeGuruInterface(int n_dims, finufft_type type, int n_vecs, BIGINT nj, FL
   
   plan.opts = opts;   /*Copy out user defined options in nufft_opts into the plan*/
   
-  int ier = make_finufft_plan(type, n_dims, n_modes, iflag, n_vecs, eps, &plan);
+  int ier = make_finufft_plan(type, n_dims, n_modes, iflag, n_transf, eps, &plan);
   if(ier){
     if(plan.opts.debug)
       printf("error (ier=%d)!\n", ier);
