+++ src/legacy/invokeGuru.h
@@ -5,7 +5,7 @@
 
 
 
-int invokeGuruInterface(int n_dims, finufft_type type, int n_vecs, BIGINT nj, FLT* xj,FLT *yj, FLT *zj, CPX* cj,int iflag,
+int invokeGuruInterface(int n_dims, finufft_type type, int n_transf, BIGINT nj, FLT* xj,FLT *yj, FLT *zj, CPX* cj,int iflag,
 			FLT eps, BIGINT *n_modes, CPX* fk, nufft_opts opts);
 
 #endif
